bugfix(UserService): update UserService methods with GlobalExceptionH…

+++ src/main/java/com/example/happy_travel/services/UserService.java
@@ -3,13 +3,12 @@ package com.example.happy_travel.services;
 import com.example.happy_travel.dtos.user.UserMapper;
 import com.example.happy_travel.dtos.user.UserRequest;
 import com.example.happy_travel.dtos.user.UserResponse;
+import com.example.happy_travel.exceptions.EntityAlreadyExistsException;
+import com.example.happy_travel.exceptions.EntityNotFoundException;
 import com.example.happy_travel.models.User;
 import com.example.happy_travel.repositories.UserRepository;
-import jakarta.validation.constraints.Null;
 import org.springframework.stereotype.Service;
 import java.util.List;
-import java.util.NoSuchElementException;
-import java.util.Optional;
 
 @Service
 public class UserService {
@@ -24,13 +23,18 @@ public class UserService {
         return users.stream().map(user -> UserMapper.toDto(user)).toList();
     }
 
-    public Optional<User> getUserById(Long id) {
-        return userRepository.findById(id);
+    public User getUserById(Long id) {
+        return userRepository.findById(id).orElseThrow(() -> new EntityNotFoundException(User.class.getSimpleName(), "Id", id.toString()));
+    }
+
+    public UserResponse getUserResponseById(Long id){
+        User user = getUserById(id);
+        return UserMapper.toDto(user);
     }
 
     public UserResponse addUser(UserRequest userRequest) {
         if (userRepository.existsByUsername(userRequest.username())){
-        throw new RuntimeException("Username already exists, please choose another");
+        throw new EntityAlreadyExistsException(User.class.getSimpleName(), "username", userRequest.username());
         }
         User newUser = UserMapper.toEntity(userRequest);
         User savedUser = userRepository.save(newUser);
@@ -38,7 +42,7 @@ public class UserService {
     }
 
     public UserResponse updateUser(Long id, UserRequest userRequest) {
-        User updatedUser = userRepository.findById(id).orElseThrow(() -> new NoSuchElementException("User not found with id " + id));
+        User updatedUser = userRepository.findById(id).orElseThrow(() -> new EntityNotFoundException(User.class.getSimpleName(), "id",id.toString()));
         updatedUser.setUsername(userRequest.username());
         updatedUser.setEmail(userRequest.email());
         updatedUser.setPassword(userRequest.password());
@@ -47,6 +51,9 @@ public class UserService {
     }
 
     public void deleteUser(Long id) {
+        if (!userRepository.existsById(id)){
+            throw new EntityNotFoundException(User.class.getSimpleName(), "id", id.toString());
+        }
         getUserById(id);
         userRepository.deleteById(id);
     }
